+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Using pseudo csv file format instead of real csv file - allows to generate smaller file sizes by using MEMORY_TOTAL only once instead in each row
 - Fixed collecting data with non integers second intervals 
 - Generated html file should be now minimized (~30% smaller)
+- Using local time offset instead of UTC time in plot
 
 ## Version 0.1.0 - 07.07.2023
 - Initial release